LayoutTests/imported/w3c/web-platform-tests/referrer-policy/unsafe-url/attr-referrer/cross-origin/http-https/script-tag/keep-origin-redirect/generic.http-expected.txt


PASS The referrer URL is stripped-referrer when a
                                 document served over http requires an https
                                 sub-resource via script-tag using the attr-referrer
                                 delivery method with keep-origin-redirect and when
                                 the target request is cross-origin.
